requirements.txt
celery>=3.1.7
django-celery>=3.1.1
six>=1.9
